refactor: reuse fields and rules from register in profile and login
