+++ src/pages/login-page/login-page.tsx
@@ -1,16 +1,23 @@
 import { useEffect } from 'react';
 import styles from './login-page.module.css';
 import { Input } from '../../components/input/input.tsx';
-import { FIELDS, validationRules } from './constants.ts';
 import { Button } from '../../components/button/button.tsx';
 import { ButtonVariants } from '../../components/button/enums.ts';
 import { messages } from './messages.ts';
 import { useForm } from 'react-hook-form';
 import { PagePath } from '../../router/enums.ts';
 import { Link, useNavigate } from 'react-router-dom';
-import type { LoginFormValues } from './types.ts';
 import { observer } from 'mobx-react-lite';
 import { userStore } from '../../store/user-store.ts';
+import {
+  FIELDS,
+  validationRules,
+} from '../../sources/constants/register-fields.ts';
+import type { RegisterFormValues } from '../../sources/types/register.ts';
+
+const LOGIN_FIELDS = FIELDS.filter(
+  field => field.name === 'password' || field.name === 'email'
+);
 
 export const LoginPage = observer(() => {
   const navigate = useNavigate();
@@ -20,9 +27,9 @@ export const LoginPage = observer(() => {
     handleSubmit,
     formState: { errors },
     reset,
-  } = useForm<LoginFormValues>();
+  } = useForm<RegisterFormValues>();
 
-  const onSubmit = async (customer: LoginFormValues) => {
+  const onSubmit = async (customer: RegisterFormValues) => {
     await userStore.login(customer);
     if (!userStore.error) {
       reset();
@@ -57,7 +64,7 @@ export const LoginPage = observer(() => {
           className={styles.formContainer}
           onSubmit={handleSubmit(onSubmit)}
         >
-          {FIELDS.map(field => {
+          {LOGIN_FIELDS.map(field => {
             const error = errors[field.name]?.message;
             const rules = validationRules[field.name];
