--- src/pages/login-page/types.ts
@@ -1,18 +1,0 @@
-export type LoginFormValues = {
-  email: string;
-  password: string;
-};
-
-export type LoginFormField = {
-  type: LoginInputFieldType;
-  label: string;
-  name: LoginFieldName;
-  placeholder?: string;
-};
-
-export enum LoginFieldName {
-  email = 'email',
-  password = 'password',
-}
-
-export type LoginInputFieldType = 'email' | 'password';
