--- src/pages/profile-page/components/general-info/constants/messages.ts
@@ -1,9 +1,0 @@
-export const messages = {
-  generalInfo: 'General Info',
-  save: 'Save',
-  edit: 'Edit',
-  dateOfBirth: 'Birth',
-  email: 'Email',
-  firstName: 'First name',
-  lastName: 'Last name',
-};
