--- src/pages/profile-page/components/general-info/form-fields.ts
@@ -1,6 +1,0 @@
-import type { Customer } from '../../../../sources/types/customer';
-
-export type FormFields = Pick<
-  Customer.Profile,
-  'email' | 'lastName' | 'firstName' | 'dateOfBirth'
->;
