+++ src/pages/profile-page/components/general-info/messages.ts
@@ -1,0 +1,5 @@
+export const messages = {
+  generalInfo: 'General Information',
+  edit: 'Edit',
+  save: 'Save',
+};
