+++ src/pages/register-page/register-page.tsx
@@ -2,12 +2,12 @@ import styles from './register-page.module.css';
 import { Button } from '../../components/button/button.tsx';
 import { ButtonVariants } from '../../components/button/enums.ts';
 import { messages } from './messages.ts';
-import { validationRules } from './constants.ts';
+import { validationRules } from '../../sources/constants/register-fields.ts';
 import { Link, useNavigate } from 'react-router-dom';
 import { PagePath } from '../../router/enums.ts';
 import { useForm } from 'react-hook-form';
-import type { RegisterFormValues } from './types.ts';
-import { RegisterFieldName } from './types.ts';
+import type { RegisterFormValues } from '../../sources/types/register.ts';
+import { RegisterFieldName } from '../../sources/types/register.ts';
 import { useState } from 'react';
 import { Checkbox } from '../../components/checkbox/checkbox.tsx';
 import { userStore } from '../../store/user-store.ts';
